Correcao Regra de negocio
