+++ PoliServices/Feature/HomeServices/HomeServicesViewController.swift
@@ -39,7 +39,7 @@ class HomeServicesViewController: UIViewController {
         homeServiceView.addGestureRecognizer(tapGesture)
         viewModel.delegate = self
         homeServiceView.subTitle.text = viewModel.dateAndHourNow()
-        checkForPermission()
+        viewModel.checkForPermission()
     }
     
     override func viewWillAppear(_ animated: Bool) {
@@ -56,51 +56,6 @@ class HomeServicesViewController: UIViewController {
     }
     
     // MARK: - Functions
-    
-    private func checkForPermission() {
-        let notificationCenter = UNUserNotificationCenter.current()
-        notificationCenter.getNotificationSettings { settings in
-            switch settings.authorizationStatus {
-            case .authorized:
-                return
-            case .denied:
-                return
-            case .notDetermined:
-                notificationCenter.requestAuthorization(options: [.alert, .sound]) { didAllow, error in
-                    if didAllow {
-                        self.dispatchNotification(hour: "", minute: "")
-                    }
-                }
-            default:
-                return
-            }
-        }
-    }
-    
-    private func dispatchNotification(hour: String, minute: String) {
-        let identifier = "my-alert"
-        let title = "Agendamento"
-        let body = "Falta 15 minutos para o serviço"
-        let hour = Int(hour)
-        let minute = Int(minute)
-        let isDaily = true
-
-        let notificationCenter = UNUserNotificationCenter.current()
-        let content = UNMutableNotificationContent()
-        content.title = title
-        content.body = body
-        content.sound = .default
-
-        let calendar = Calendar.current
-        var dateComponents = DateComponents(calendar: calendar, timeZone: TimeZone.current)
-        dateComponents.hour = hour
-        dateComponents.minute = minute
-
-        let trigger = UNCalendarNotificationTrigger(dateMatching: dateComponents, repeats: isDaily)
-        let request = UNNotificationRequest(identifier: identifier, content: content, trigger: trigger)
-        notificationCenter.removePendingNotificationRequests(withIdentifiers: [identifier])
-        notificationCenter.add(request)
-    }
     
     @objc func handleTap() {
         let serviceDetailViewModel = ServiceDetailViewModel()
@@ -126,10 +81,6 @@ extension HomeServicesViewController: HomeServicesViewDelegate {
 // MARK: - HomeServicesViewModelDelegate
 
 extension HomeServicesViewController: HomeServicesViewModelDelegate {
-    func hourAndMinuteNotice(hour: String, minute: String) {
-        dispatchNotification(hour: hour, minute: minute)
-    }
-    
     func onStartDataHomeService(model: HomeModel) {
         homeServiceView.setParameter(model: model)
     }
